feat: file 부분 컴포넌트로 분리
