+++ src/app/_components/sidebar/finder/finder-card.tsx
@@ -1,6 +1,6 @@
 import { css } from '@/../styled-system/css';
 
-import PageCloseIcon from '@/icons/page-close-icon';
+import File from './file';
 import fileList from './file-list-mock';
 
 const finderCard = css({
@@ -16,44 +16,11 @@ const finderCard = css({
   boxShadow: 'sidebar',
 });
 
-const fileStyle = css({
-  display: 'flex',
-  flexDirection: 'row',
-  alignItems: 'center',
-  height: '2rem',
-  gap: 'tiny',
-  fontWeight: 'regular',
-  fontSize: 'md',
-  borderRadius: '0.25rem',
-  cursor: 'pointer',
-  px: 'small',
-
-  '&:hover': {
-    backgroundColor: '#F1F1F0',
-  },
-});
-
-const fileToggleButton = css({
-  borderRadius: '0.25rem',
-
-  '&:hover': {
-    backgroundColor: '#E4E4E3',
-  },
-});
-
 const FinderCard = () => {
   return (
     <div className={finderCard}>
       {fileList.map(file => {
-        return (
-          <div className={fileStyle}>
-            <div className={fileToggleButton}>
-              <PageCloseIcon color="black" />
-            </div>
-            <div>{file.icon}</div>
-            <div>{file.title}</div>
-          </div>
-        );
+        return <File key={file.docsId} file={file} />;
       })}
     </div>
   );
