+++ src/types/file-type.ts
@@ -1,0 +1,8 @@
+interface IFileType {
+  docsId: number;
+  title: string;
+  icon: string;
+  children: IFileType[];
+}
+
+export default IFileType;
